Allow a dependency to have no files
The schema requires at least one file to be pulled in when declaring a dependency.

This requirement seems unnecessary. Having no dependencies is probably not very useful in practice, but it can help make integration tests that aren't concerned with file copying more concise. It also seems harmless to allow this scenario.