Restrict the syntax of relative paths. (#5)
Paths contained in a packet metadata need to be relative and as
cross-platform as possible. In particular, we've settled on using
forward slashes everywhere, even for packets produced on Windows.

While implementations should generally catch such invalid paths earlier
on, having the restriction at the schema level will help catch anything
that slipped through.

The regex used is roughly based on the [Windows documentation][0]. It
isn't intended as a fully comprehensive list, eg. it does not reject the
filenames `NUL` or `COM1`, nor does it forbid paths that contain `.` or
`..`. POSIX allows anything except NULL bytes.

I also noticed the CI script was't effective: it was catching all
exceptions, printing them and returning successfully always. I've fixed
it by exiting with code 1 if any exception is raised, and also added a
bit of validations to the schemas themselves.

[0]: https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file